feat: Create IAM user for Grafana to use to read CloudWatch [DEV-164] (#6)
## What is the motivation behind this PR?
We are looking to use Grafana as our single pane of glass into the
greater over-all system, including AWS. As such we need the ability to
view CloudWatch metrics from Grafana, which means we need credentials to
do so. It was decided to use a IAM user and manually provision the
source, as opposed to an IAM role on the EC2 instance, as we don't want
any pod in the cluster to be able to pull from cloudwatch.

## Summary of changes
* Create IAM user and store in SecretsManager.

## Author reminders
- [x] Reviewed my own PR
- [ ] Updated documentation, if applicable